+++ package.json
@@ -7,6 +7,8 @@
     "babel-loader": "^7.1.1",
     "babel-preset-env": "^1.6.0",
     "babel-preset-flow": "^6.23.0",
+    "clean-webpack-plugin": "^0.1.16",
+    "copy-webpack-plugin": "^4.0.1",
     "eslint": "^3.19.0",
     "eslint-config-airbnb-base": "^11.2.0",
     "eslint-config-prettier": "^2.3.0",
@@ -15,23 +17,25 @@
     "eslint-plugin-import": "^2.7.0",
     "eslint-plugin-prettier": "^2.1.2",
     "husky": "^0.14.3",
+    "imports-loader": "^0.7.1",
     "lint-staged": "^4.0.2",
     "prettier-eslint-cli": "^4.1.1",
-    "rimraf": "^2.6.1",
     "webpack": "^3.3.0"
   },
   "scripts": {
     "lint": "eslint src",
-    "prebuild": "yarn lint && yarn clear",
+    "prebuild": "yarn lint",
     "build": "webpack",
     "build:watch": "webpack -w",
     "precommit": "yarn lint-staged",
     "predeploy": "yarn build",
-    "deploy": "gapps upload",
-    "clear": "rimraf build/**"
+    "deploy": "gapps upload"
   },
   "lint-staged": {
-    "*.js": ["prettier-eslint \"**/*.js\" --write", "git add"]
+    "*.js": [
+      "prettier-eslint \"**/*.js\" --write",
+      "git add"
+    ]
   },
   "dependencies": {}
 }
