+++ src/Code.js
@@ -1,0 +1,10 @@
+/* global app */
+
+function onOpen() {
+  app.doOnOpen();
+}
+
+// eslint-disable-next-line no-unused-vars
+function onInstall() {
+  onOpen();
+}
