+++ src/makeFormOffline.js
@@ -1,0 +1,5 @@
+// @flow
+
+export default function makeFormOffline() {
+  FormApp.getUi().alert('Hello from Forms Offline Addon!');
+}
