+++ test.js
@@ -1,0 +1,4 @@
+/* eslint-disable */
+function onOpen() {
+  console.log('test');
+}
